+++ packages/contracts/src/systems/PrizeDistributionSystem.sol
@@ -19,22 +19,20 @@ import {WorldContextConsumer, WORLD_CONTEXT_CONSUMER_INTERFACE_ID} from "@lattic
 //   uint256 burnedAmount;
 // }
 
-
 contract PrizeDistributionSystem is System {
-
     // return true if the game is over
     // return false if the game is not over
     function prizeDistribute() public payable returns (bool) {
         GameData memory gameData = Game.get();
-        uint start_time = gameData.startTime;
-        uint current_round = gameData.currentRound;
-        uint penultimate_deadline = start_time + (current_round - 2) * 1 days;
-        uint last_deadline = start_time + (current_round - 1) * 1 days;
+        uint256 start_time = gameData.startTime;
+        uint256 current_round = gameData.currentRound;
+        uint256 penultimate_deadline = start_time + (current_round - 2) * 1 days;
+        uint256 last_deadline = start_time + (current_round - 1) * 1 days;
         address[] memory all_players = gameData.allPlayers;
 
-        uint numAlives = 0;
-        uint numLastAlives = 0;
-        
+        uint256 numAlives = 0;
+        uint256 numLastAlives = 0;
+
         uint256 prize = address(this).balance;
 
         // check the number of players
@@ -47,8 +45,7 @@ contract PrizeDistributionSystem is System {
                 numAlives++;
             }
 
-            if (playerData.lastCheckedTime > penultimate_deadline
-                    && playerData.lastCheckedTime < last_deadline) {
+            if (playerData.lastCheckedTime > penultimate_deadline && playerData.lastCheckedTime < last_deadline) {
                 numLastAlives++;
             }
         }
@@ -68,7 +65,7 @@ contract PrizeDistributionSystem is System {
 
         // if there is no player,
         // distribute the prize to the players who were alive in last round
-        
+
         if (numAlives == 0) {
             for (uint32 i = 0; i < all_players.length; i++) {
                 address playerAddr = all_players[i];
@@ -79,8 +76,7 @@ contract PrizeDistributionSystem is System {
                     payable(playerAddr).transfer(address(this).balance);
                 }
 
-                if (playerData.lastCheckedTime > penultimate_deadline
-                        && playerData.lastCheckedTime < last_deadline) {
+                if (playerData.lastCheckedTime > penultimate_deadline && playerData.lastCheckedTime < last_deadline) {
                     payable(playerAddr).transfer(prize / numLastAlives);
                 }
             }
